+++ marf/src/app/calculators/dot-product/dot-product-routing.module.ts
@@ -1,8 +1,33 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { PageNotFoundComponent } from 'src/app/page-not-found/page-not-found.component';
+import { DotProductComponent } from './dot-product.component';
 
-
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: '', 
+    component: DotProductComponent, 
+    children: [
+      { 
+        path: 'interactive', 
+        component: DotProductComponent 
+      },
+      // {
+      //   path: 'examples',
+      //   // TODO: Add component or template or something
+      // },
+      {
+        path: '',
+        redirectTo: 'interactive',
+        pathMatch: 'full'
+      },
+      {
+        path: '**',
+        component: PageNotFoundComponent
+      }
+    ]
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
