+++ marf/src/app/calculators/dot-product/dot-product.module.ts
@@ -1,40 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { RouterModule, Routes } from '@angular/router';
-import { DotProductComponent } from './dot-product.component';
-import { PageNotFoundComponent } from 'src/app/page-not-found/page-not-found.component';
-
-const routes: Routes = [
-  {
-    path: '', 
-    component: DotProductComponent, 
-    children: [
-      { 
-        path: 'interactive', 
-        component: DotProductComponent 
-      },
-      // {
-      //   path: 'examples',
-      //   // TODO: Add component or template or something
-      // },
-      {
-        path: '',
-        redirectTo: 'interactive',
-        pathMatch: 'full'
-      },
-      {
-        path: '**',
-        component: PageNotFoundComponent
-      }
-    ]
-  }
-];
+import { DotProductRoutingModule } from './dot-product-routing.module';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   imports: [
     CommonModule,
-    RouterModule.forChild(routes)
+    DotProductRoutingModule
   ],
   exports: [RouterModule]
 })
