+++ marf/src/app/calculators/prime-numbers/prime-numbers.module.ts
@@ -1,40 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { PrimeNumbersComponent } from './prime-numbers.component';
-import { RouterModule, Routes } from '@angular/router';
-import { PageNotFoundComponent } from 'src/app/page-not-found/page-not-found.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: PrimeNumbersComponent,
-    children: [
-      {
-        path: 'interactive',
-        component: PrimeNumbersComponent
-      },
-      // {
-      //   path: 'examples',
-      //   // TODO: Add component or template or something
-      // },
-      {
-        path: '',
-        redirectTo: 'interactive',
-        pathMatch: 'full'
-      },
-      {
-        path: '**',
-        component: PageNotFoundComponent
-      }
-    ]
-  }
-];
+import { PrimeNumbersRoutingModule } from './prime-numbers-routing.module';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   imports: [
     CommonModule,
-    RouterModule.forChild(routes)
+    PrimeNumbersRoutingModule
   ],
   exports: [RouterModule]
 })
